+++ src/main/java/org/dcsa/standards/specifications/standards/tnt/v300/model/ReeferDetails.java
@@ -3,10 +3,13 @@ package org.dcsa.standards.specifications.standards.tnt.v300.model;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import lombok.Data;
+import org.dcsa.standards.specifications.standards.tnt.v300.types.ReeferPowerStatusCode;
 
 @Schema(description = "Reefer-specific details")
 @Data
 public class ReeferDetails {
+
+  @Schema protected ReeferPowerStatusCode powerStatus;
 
   @Schema(
       description =